refactor(skill-gate-eval): address PR feedback
- Pull all shared types into types.ts; PreferenceValue is now a string
  enum, not a union literal.
- Pull module-level constants into constants.ts (ASK_QUESTION_TOOL,
  PASS_THRESHOLD, DEFAULT_MODEL).
- scenario.ts becomes IO-only; harness.ts and run.ts import from
  types/constants instead of redeclaring inline.
- buildSystemPrompt's docstring now shows the literal output shape.
- Verdict return uses long-form property assignments.
- Drop dated design-doc references in code comments and READMEs;
  point at the architecture/ folder by topic instead so the refs
  don't rot when the doc is renamed.
- Add internal/skill-gate-eval/tsconfig.json so the harness sources
  typecheck independently of the main project build.

Layer 1 lint still green: 49/49.
