+++ internal/skill-gate-eval/src/constants.ts
@@ -1,0 +1,8 @@
+/** Tool name the agent SDK uses for its built-in user-prompt tool. */
+export const ASK_QUESTION_TOOL = "AskQuestion";
+
+/** Per-scenario pass-rate floor below which we treat the gate as misfiring. */
+export const PASS_THRESHOLD = 0.99;
+
+/** Default eval target until the suite is stable enough to add others. */
+export const DEFAULT_MODEL = "claude-sonnet-4-6";
